+++ ceammc/CHANGELOG.md
@@ -10,6 +10,7 @@
   - @norm property added to ui.env to do envelope value normalization on input
   - @value property added to seq.counter
   - @dev property added to flow.space: to set deviation
+  - @channels property added to xdac~: live-channel remapping
 - new methods:
   - [polltime( method added to system.cursor to change global polltime
 - misc: